+++ scripts/RideableSettings.js
@@ -1,20 +1,30 @@
+import { RideableCompUtils, cWallHeight } from "./RideableCompUtils.js";
 import { RideableUtils, cModuleName, Translate} from "./RideableUtils.js";
 import { MountSelected, MountSelectedFamiliar, UnMountSelected } from "./MountingScript.js";
 
 Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   //Settings
   game.settings.register(cModuleName, "RidingHeight", {
-	name: Translate("RidingHeight.name"),
-	hint: Translate("RidingHeight.descrp"),
+	name: Translate("Settings.RidingHeight.name"),
+	hint: Translate("Settings.RidingHeight.descrp"),
 	scope: "world",
 	config: true,
 	type: Number,
 	default: "5"
   });
+  
+  game.settings.register(cModuleName, "useRiddenTokenHeight", {
+	name: Translate("Settings.useRiddenTokenHeight.name"),
+	hint: Translate("Settings.useRiddenTokenHeight.descrp"),
+	scope: "world",
+	config: RideableCompUtils.isactiveModule(cWallHeight),
+	type: Boolean,
+	default: true
+  });
 
   game.settings.register(cModuleName, "MountingDistance", {
-	name: Translate("MountingDistance.name"),
-	hint: Translate("MountingDistance.descrp"),
+	name: Translate("Settings.MountingDistance.name"),
+	hint: Translate("Settings.MountingDistance.descrp"),
 	scope: "world",
 	config: true,
 	type: Number,
@@ -22,8 +32,8 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   });
   
   game.settings.register(cModuleName, "BorderDistance", {
-	name: Translate("BorderDistance.name"),
-	hint: Translate("BorderDistance.descrp"),
+	name: Translate("Settings.BorderDistance.name"),
+	hint: Translate("Settings.BorderDistance.descrp"),
 	scope: "world",
 	config: true,
 	type: Boolean,
@@ -31,8 +41,8 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   });
   
   game.settings.register(cModuleName, "MaxRiders", {
-	name: Translate("MaxRiders.name"),
-	hint: Translate("MaxRiders.descrp"),
+	name: Translate("Settings.MaxRiders.name"),
+	hint: Translate("Settings.MaxRiders.descrp"),
 	scope: "world",
 	config: true,
 	type: Number,
@@ -40,22 +50,31 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   });
   
   game.settings.register(cModuleName, "RiderMovement", {
-	name: Translate("RiderMovement.name"),
-	hint: Translate("RiderMovement.descrp"),
+	name: Translate("Settings.RiderMovement.name"),
+	hint: Translate("Settings.RiderMovement.descrp"),
 	scope: "world",
 	config: true,
 	type: String,
 	choices: {
-		"RiderMovement-disallow": Translate("RiderMovement.RiderMovement-disallow"),
-		"RiderMovement-dismount": Translate("RiderMovement.RiderMovement-dismount"),
-		"RiderMovement-moveridden": Translate("RiderMovement.RiderMovement-moveridden")
+		"RiderMovement-disallow": Translate("Settings.RiderMovement.options.disallow"),
+		"RiderMovement-dismount": Translate("Settings.RiderMovement.options.dismount"),
+		"RiderMovement-moveridden": Translate("Settings.RiderMovement.options.moveridden")
 	},
 	default: "RiderMovement-disallow"
   });
   
+  game.settings.register(cModuleName, "RiderRotation", {
+	name: Translate("Settings.RiderRotation.name"),
+	hint: Translate("Settings.RiderRotation.descrp"),
+	scope: "world",
+	config: true,
+	type: Boolean,
+	default: false
+  });   
+  
   game.settings.register(cModuleName, "RidingSystemEffects", {
-	name: Translate("RidingSystemEffects.name"),
-	hint: Translate("RidingSystemEffects.descrp"),
+	name: Translate("Settings.RidingSystemEffects.name"),
+	hint: Translate("Settings.RidingSystemEffects.descrp"),
 	scope: "world",
 	config: true,
 	type: Boolean,
@@ -63,8 +82,8 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   });  
   
   game.settings.register(cModuleName, "RideableTag", {
-	name: Translate("RideableTag.name"),
-	hint: Translate("RideableTag.descrp"),
+	name: Translate("Settings.RideableTag.name"),
+	hint: Translate("Settings.RideableTag.descrp"),
 	scope: "world",
 	config: RideableUtils.isPf2e(),
 	type: Boolean,
@@ -72,8 +91,8 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   });  
   
   game.settings.register(cModuleName, "FamiliarRiding", {
-	name: Translate("FamiliarRiding.name"),
-	hint: Translate("FamiliarRiding.descrp"),
+	name: Translate("Settings.FamiliarRiding.name"),
+	hint: Translate("Settings.FamiliarRiding.descrp"),
 	scope: "world",
 	config: true,
 	type: Boolean,
@@ -81,8 +100,8 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   });  
   
   game.settings.register(cModuleName, "PreventEnemyRiding", {
-	name: Translate("PreventEnemyRiding.name"),
-	hint: Translate("PreventEnemyRiding.descrp"),
+	name: Translate("Settings.PreventEnemyRiding.name"),
+	hint: Translate("Settings.PreventEnemyRiding.descrp"),
 	scope: "world",
 	config: true,
 	type: Boolean,
@@ -90,8 +109,8 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   });  
   
   game.settings.register(cModuleName, "MessagePopUps", {
-	name: Translate("MessagePopUps.name"),
-	hint: Translate("MessagePopUps.descrp"),
+	name: Translate("Settings.MessagePopUps.name"),
+	hint: Translate("Settings.MessagePopUps.descrp"),
 	scope: "client",
 	config: true,
 	type: Boolean,
@@ -100,8 +119,8 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   
   //Keys
   game.keybindings.register(cModuleName, "Mount", {
-    name: Translate("Mount.name"),
-    hint: Translate("Mount.descrp"),
+    name: Translate("Keys.Mount.name"),
+    hint: Translate("Keys.Mount.descrp"),
     editable: [
       {
         key: "KeyM"
@@ -113,8 +132,8 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   });
   
   game.keybindings.register(cModuleName, "UnMount", {
-    name: Translate("UnMount.name"),
-    hint: Translate("UnMount.descrp"),
+    name: Translate("Keys.UnMount.name"),
+    hint: Translate("Keys.UnMount.descrp"),
     editable: [
       {
         key: "KeyN"
@@ -126,8 +145,8 @@ Hooks.once("init", () => {  // game.settings.get(cModuleName, "")
   });
   
   game.keybindings.register(cModuleName, "MountFamiliar", {
-    name: Translate("MountFamiliar.name"),
-    hint: Translate("MountFamiliar.descrp"),
+    name: Translate("Keys.MountFamiliar.name"),
+    hint: Translate("Keys.MountFamiliar.descrp"),
     editable: [
       {
         key: "KeyJ"